Tidy up Substance overrides.
