+++ src/kit/ui/index.js
@@ -2,9 +2,13 @@ export { default as AbstractScrollPane } from './AbstractScrollPane'
 export { default as BodyScrollPane } from './BodyScrollPane'
 export { default as Button } from './Button'
 export { default as CheckboxInput } from './CheckboxInput'
+export { default as Clipboard } from './_Clipboard'
 export * from './componentHelpers'
+export { default as ContainerEditor } from './_ContainerEditor'
 export { default as ContextMenu } from './ContextMenu'
 export { default as FormRowComponent } from './FormRowComponent'
+export { default as IsolatedInlineNodeComponent } from './_IsolatedInlineNodeComponent'
+export { default as IsolatedNodeComponent } from './_IsolatedNodeComponent'
 export { default as Managed } from './Managed'
 export { default as Menu } from './Menu'
 export { default as MenuGroup } from './MenuGroup'
@@ -15,6 +19,10 @@ export { default as Overlay } from './Overlay'
 export { default as OverlayMixin } from './OverlayMixin'
 export { default as PinnedMessage } from './PinnedMessage'
 export { default as ScrollPane } from './ScrollPane'
+export { default as Surface } from './_Surface'
+export { default as TextInput } from './TextInput'
+export { default as TextPropertyComponent } from './_TextPropertyComponent'
+export { default as TextPropertyEditor } from './_TextPropertyEditor'
 export { default as ToggleTool } from './ToggleTool'
 export { default as Toolbar } from './Toolbar'
 export { default as ToolDropdown } from './ToolDropdown'
